WebSocket_Client_Server/requirements.txt
asyncio
websockets
pycryptodome